+++ CHANGELOG.md
@@ -9,6 +9,7 @@ New features:
 
 Bug fixes:
 
+* `ENV` now reads the process environment directly, so variables set by native code, such as a C extension calling `setenv()`, are visible (@nirvdrum).
 * Raise a `RuntimeError` when using an object from a `Polyglot::InnerContext` after it has been closed (#4061, @eregon).
 * Fix `object.method(:send).to_proc.call(some_method)` (#4299, @eregon).
 * Fix `TruffleRuby::AtomicReference#compare_and_set` to work with NaN values like `Float::NAN` (@eregon).